+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Shell: add 3 different instanciation methods, 
 - Shell constructors: make_extrusion, make_wedge
 - Solid constructors: make_extrusion, make_extrusion_from_frame_and_wires, make_wedge
+- Shape: to_dict, dict_to_object, to_brep, from_brep, from_brep_stream, to_brep_stream, 
+
 
 
 #### core.py